add default.go & update snapshots
