+++ tests/snapshots/TestCLICommands_atmos_describe_config.stdout.golden
@@ -145,5 +145,5 @@
       "color": true
     }
   },
-  "cli_config_path": "/absolute/path/to/repo/examples/demo-stacks/atmos"
+  "cli_config_path": "/absolute/path/to/repo/examples/demo-stacks/atmos.yaml"
 }
