+++ tests/snapshots/TestCLICommands_atmos_describe_config_-f_yaml.stdout.golden
@@ -43,5 +43,5 @@ default: false
 validate:
     editorconfig:
         color: true
-cli_config_path: /absolute/path/to/repo/examples/demo-stacks/atmos
+cli_config_path: /absolute/path/to/repo/examples/demo-stacks/atmos.yaml
