+++ tests/test-cases/env.yaml
@@ -16,7 +16,7 @@ tests:
       diff: []
       stdout:
         - 'atmos_base_path = "./"'
-        - 'atmos_cli_config_path = ".*tests.*fixtures.*scenarios.*env.*atmos"'
+        - 'atmos_cli_config_path = ".*tests.*fixtures.*scenarios.*env.*atmos.yaml"'
       stderr:
         - "^$"
       exit_code: 0
